#1: Migrated to "busted" framework for testing.
