--- .buildpath
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<buildpath>
-	<buildpathentry kind="src" path="src/main/lua"/>
-	<buildpathentry kind="src" path="src/test/lua"/>
-	<buildpathentry kind="con" path="org.eclipse.ldt.ExecutionEnvironmentContainer/lua/5.1"/>
-</buildpath>
