+++ .gitignore
@@ -1,5 +1,3 @@
-# This is a hybrid Lua /Java project, so we need to check in .project and .classpath
-
 /target/
 *.log
 *~
