--- .settings/org.eclipse.core.resources.prefs
@@ -1,5 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/main/lua=UTF-8
-encoding//src/test/lua=UTF-8
-encoding//src/test/resources=UTF-8
-encoding/<project>=UTF-8
