--- .settings/org.eclipse.ldt.prefs
@@ -1,2 +1,0 @@
-Grammar__default_id=lua-5.1
-eclipse.preferences.version=1
