+++ dependencies.md
@@ -1,87 +1,26 @@
 <!-- @formatter:off -->
 # Dependencies
 
-## Compile Dependencies
+`virtual-schema-common-lua` is a base library. It is intended to be used as a dependency in concrete implementations of Exasol Virtual Schemas.
 
-| Dependency                        | License  |
-| --------------------------------- | -------- |
-| [Maven Project Version Getter][0] | [MIT][1] |
+To use this library at runtime you need Lua 5.4 or later.
+
+## Runtime Dependencies
+
+| Dependency                                             | License     |
+|--------------------------------------------------------|-------------|
+| [Lua 5.4](https://www.lua.org)                         | [MIT][mit]  |
+| [lua-cjson](https://github.com/openresty/lua-cjson)    | [MIT][mit]  |
+| [remotelog](https://github.com/exasol/remotelog-lua)   | [MIT][mit]  |
+
+`lua-cjson` depends on the `cjson` library both are preinstalled on Exasol.
+
+`remotelog` depends on `luasocket` which is also preinstalled.
 
 ## Test Dependencies
 
-| Dependency                                     | License                           |
-| ---------------------------------------------- | --------------------------------- |
-| [EXASolution JDBC Driver][2]                   | [EXAClient License][3]            |
-| [Test containers for Exasol on Docker][4]      | [MIT][1]                          |
-| [Testcontainers :: JUnit Jupiter Extension][6] | [MIT][7]                          |
-| [Hamcrest][8]                                  | [BSD License 3][9]                |
-| [Matcher for SQL Result Sets][10]              | [MIT][1]                          |
-| [JUnit Jupiter Engine][12]                     | [Eclipse Public License v2.0][13] |
-| [JUnit Platform Runner][12]                    | [Eclipse Public License v2.0][13] |
-| [JUnit Jupiter Params][12]                     | [Eclipse Public License v2.0][13] |
-| [SLF4J JDK14 Binding][18]                      | [MIT License][19]                 |
-| [Test Database Builder for Java][20]           | [MIT][1]                          |
+| Dependency                               | License    |
+|------------------------------------------|------------|
+| [busted](http://olivinelabs.com/busted/) | [MIT][mit] |
 
-## Plugin Dependencies
-
-| Dependency                                              | License                                        |
-| ------------------------------------------------------- | ---------------------------------------------- |
-| [Project keeper maven plugin][22]                       | [MIT][1]                                       |
-| [Apache Maven Enforcer Plugin][24]                      | [Apache License, Version 2.0][25]              |
-| [Maven Surefire Plugin][26]                             | [Apache License, Version 2.0][25]              |
-| [Apache Maven Compiler Plugin][28]                      | [Apache License, Version 2.0][25]              |
-| [Maven Failsafe Plugin][30]                             | [Apache License, Version 2.0][25]              |
-| [Exec Maven Plugin][32]                                 | [Apache License 2][33]                         |
-| [org.sonatype.ossindex.maven:ossindex-maven-plugin][34] | [ASL2][33]                                     |
-| [OpenFastTrace Maven Plugin][36]                        | [GNU General Public License v3.0][37]          |
-| [Build Helper Maven Plugin][38]                         | [The MIT License][39]                          |
-| [Versions Maven Plugin][40]                             | [Apache License, Version 2.0][25]              |
-| [JaCoCo :: Maven Plugin][42]                            | [Eclipse Public License 2.0][43]               |
-| [error-code-crawler-maven-plugin][44]                   | [MIT][1]                                       |
-| [Apache Maven JAR Plugin][46]                           | [Apache License, Version 2.0][25]              |
-| [Reproducible Build Maven Plugin][48]                   | [Apache 2.0][33]                               |
-| [Maven Clean Plugin][50]                                | [The Apache Software License, Version 2.0][33] |
-| [Maven Resources Plugin][52]                            | [The Apache Software License, Version 2.0][33] |
-| [Maven Install Plugin][54]                              | [The Apache Software License, Version 2.0][33] |
-| [Maven Deploy Plugin][56]                               | [The Apache Software License, Version 2.0][33] |
-| [Maven Site Plugin 3][58]                               | [The Apache Software License, Version 2.0][33] |
-
-[22]: https://github.com/exasol/project-keeper-maven-plugin
-[42]: https://www.eclemma.org/jacoco/index.html
-[33]: http://www.apache.org/licenses/LICENSE-2.0.txt
-[26]: https://maven.apache.org/surefire/maven-surefire-plugin/
-[50]: http://maven.apache.org/plugins/maven-clean-plugin/
-[3]: https://docs.exasol.com/connect_exasol/drivers/jdbc.htm
-[1]: https://opensource.org/licenses/MIT
-[30]: https://maven.apache.org/surefire/maven-failsafe-plugin/
-[20]: https://github.com/exasol/test-db-builder-java
-[0]: https://github.com/exasol/maven-project-version-getter
-[32]: http://www.mojohaus.org/exec-maven-plugin
-[40]: http://www.mojohaus.org/versions-maven-plugin/
-[9]: http://opensource.org/licenses/BSD-3-Clause
-[28]: https://maven.apache.org/plugins/maven-compiler-plugin/
-[7]: http://opensource.org/licenses/MIT
-[36]: https://github.com/itsallcode/openfasttrace-maven-plugin
-[43]: https://www.eclipse.org/legal/epl-2.0/
-[4]: https://github.com/exasol/exasol-testcontainers
-[10]: https://github.com/exasol/hamcrest-resultset-matcher
-[48]: http://zlika.github.io/reproducible-build-maven-plugin
-[37]: https://www.gnu.org/licenses/gpl-3.0.html
-[39]: https://opensource.org/licenses/mit-license.php
-[19]: http://www.opensource.org/licenses/mit-license.php
-[25]: https://www.apache.org/licenses/LICENSE-2.0.txt
-[24]: https://maven.apache.org/enforcer/maven-enforcer-plugin/
-[2]: http://www.exasol.com
-[13]: https://www.eclipse.org/legal/epl-v20.html
-[54]: http://maven.apache.org/plugins/maven-install-plugin/
-[12]: https://junit.org/junit5/
-[34]: https://sonatype.github.io/ossindex-maven/maven-plugin/
-[6]: https://testcontainers.org
-[38]: http://www.mojohaus.org/build-helper-maven-plugin/
-[8]: http://hamcrest.org/JavaHamcrest/
-[18]: http://www.slf4j.org
-[56]: http://maven.apache.org/plugins/maven-deploy-plugin/
-[58]: http://maven.apache.org/plugins/maven-site-plugin/
-[52]: http://maven.apache.org/plugins/maven-resources-plugin/
-[44]: https://github.com/exasol/error-code-crawler-maven-plugin
-[46]: https://maven.apache.org/plugins/maven-jar-plugin/
+[mit]: https://mit-license.org/
